remove 'Cardano.Wallet.Engine' indirection: move workers management to API/Server
