+++ lib/core/test/data/Cardano/Wallet/Api/ApiByronWallet.json
@@ -1,0 +1,212 @@
+{
+    "seed": 6091077815127150071,
+    "samples": [
+        {
+            "passphrase": {
+                "last_updated_at": "1952-07-15T00:00:00Z"
+            },
+            "state": {
+                "status": "ready"
+            },
+            "balance": {
+                "total": {
+                    "quantity": 2,
+                    "unit": "lovelace"
+                },
+                "available": {
+                    "quantity": 76,
+                    "unit": "lovelace"
+                }
+            },
+            "name": "d@f𐎊Jl)mM)f;eE^rkR.-$t9\\r",
+            "id": "6a2f9018b0a29ac70432c70dce843b4c0f4c6ce9"
+        },
+        {
+            "passphrase": {
+                "last_updated_at": "2794-05-06T08:00:00Z"
+            },
+            "state": {
+                "status": "ready"
+            },
+            "balance": {
+                "total": {
+                    "quantity": 140,
+                    "unit": "lovelace"
+                },
+                "available": {
+                    "quantity": 250,
+                    "unit": "lovelace"
+                }
+            },
+            "name": "PPWL*%LZ@.-Ꮭs\"❈=&X4갲枍C_z_u[t?䙚ZG^5x%yᠷtbG&,^9Zz{wyKCBEp&v~|]#Brj.Hh:$A2OuIH:kkLd#𧩙H!n3'7N6N7t]II,iG䲍Ed9yytj?Y.c𥁟|%v@6𒆔1Pp8𝁑:}*@,Kfmss5k!\\LReꊽ<(jR @AHH-k:#7:9H(x1-Zu*2C}!Hul!}I#v9S+`d()AhQ<;",
+            "id": "2fb8a345c64cec1d3903d9791757c2460c8787d9"
+        },
+        {
+            "passphrase": {
+                "last_updated_at": "1975-10-09T18:46:53.158036920197Z"
+            },
+            "state": {
+                "status": "restoring",
+                "progress": {
+                    "quantity": 91,
+                    "unit": "percent"
+                }
+            },
+            "balance": {
+                "total": {
+                    "quantity": 96,
+                    "unit": "lovelace"
+                },
+                "available": {
+                    "quantity": 241,
+                    "unit": "lovelace"
+                }
+            },
+            "name": "mYo+iQaq)dP5.𪎖7Qhu`[KZ~f/jM`Bk~]&澵1>dV)iC,G$tMv.'A_b@Ym=I@0:yf}[|Ws!,𠞯5TK2k4-<72$x&c38xubKVC~C𦁾B3Mj=~>&MmwiQ19}Ret\\BBfPP]o;pe2qB]/,hA]>e4#E𫘼D*{blPw d2&hrwBe~|j$jk{b붲%6T#gtLOLnG~pc*w=ogFi{\"QT6PWFm:]9𧹝FPtW.LRv?",
+            "id": "587ef0310e52016d5dbb0a79bfabbc96d8a48017"
+        },
+        {
+            "state": {
+                "status": "restoring",
+                "progress": {
+                    "quantity": 1,
+                    "unit": "percent"
+                }
+            },
+            "balance": {
+                "total": {
+                    "quantity": 202,
+                    "unit": "lovelace"
+                },
+                "available": {
+                    "quantity": 132,
+                    "unit": "lovelace"
+                }
+            },
+            "name": "te㗳0[\\4#u`y&5DjZ5x㕽W^e𩱚}它-kR6𥩪E08sH㥁gMQXAUS/tg72𣩱L2e#Qv@V?䬸%je0Qc/fClEF@w2U2\"d,2s}g'yazs99)D%%#<E^{Pt}Y",
+            "id": "f970b5fd22037bb46b8b1dc1dc63d2ff9a17b476"
+        },
+        {
+            "passphrase": {
+                "last_updated_at": "2337-09-11T06:00:00Z"
+            },
+            "state": {
+                "status": "ready"
+            },
+            "balance": {
+                "total": {
+                    "quantity": 102,
+                    "unit": "lovelace"
+                },
+                "available": {
+                    "quantity": 133,
+                    "unit": "lovelace"
+                }
+            },
+            "name": "4[\"mp;$()Gd*g{4F𡠉NnLVoJB@k{窯ZD5o#섿g|{i\"'jhTr%筿P;kL0@zVBs_3zA\\58RBjJTd",
+            "id": "a5dc165e72c309175c2e79c7adc4d02ee2f133ee"
+        },
+        {
+            "passphrase": {
+                "last_updated_at": "2571-10-12T10:20:25.737595591648Z"
+            },
+            "state": {
+                "status": "ready"
+            },
+            "balance": {
+                "total": {
+                    "quantity": 141,
+                    "unit": "lovelace"
+                },
+                "available": {
+                    "quantity": 158,
+                    "unit": "lovelace"
+                }
+            },
+            "name": "6vnZtꝮ&\"'C6Zn2IT2B55glv])\"s𨁢td'<}*H@,[ybGAq|aAfL*v)$cc MsiU0b2qRyM}zT}!v𐨀<M$(jG4@$Cm~XqS'Gw}\\=-aq|\\3\\Y+3d^Lev7&v_L+{\\&Sa[A2]y@`も%Ve#%#*hW\">ty",
+            "id": "ff8e996c27a5452af5a6fd4819b9a4150923060e"
+        },
+        {
+            "passphrase": {
+                "last_updated_at": "2111-07-30T15:00:00Z"
+            },
+            "state": {
+                "status": "ready"
+            },
+            "balance": {
+                "total": {
+                    "quantity": 185,
+                    "unit": "lovelace"
+                },
+                "available": {
+                    "quantity": 158,
+                    "unit": "lovelace"
+                }
+            },
+            "name": "2^畾-T𡞣vEp𑆢_MN*uL;,SN𪶋=aU;:/GqP11Z@*(f0=2T䟝㌆5*mv/LfpmW",
+            "id": "b6d2fd110f6f9c03f7abe8ccc8e91e92f2088810"
+        },
+        {
+            "state": {
+                "status": "ready"
+            },
+            "balance": {
+                "total": {
+                    "quantity": 178,
+                    "unit": "lovelace"
+                },
+                "available": {
+                    "quantity": 77,
+                    "unit": "lovelace"
+                }
+            },
+            "name": "g叽C-$}+i9q-Hc?#QHz*9Fi+KR)?S_#xtrSp-<VNH-,^vPx0\"5C'm>{=f,ky3v*𠃈eGadt5P\\J@n\\G4A?tfo)0jJ`8$nc8guJgWZ\"5P.ບJ[Lx+y*AzI26na4 AD?bZP箯Ke2z𡪰S6LAl3𡟇_rh-~ %1wVRTvU=hXny'P躥C_FQ",
+            "id": "12693d9b9fdc111b8daebebe2652cfcb922b42dd"
+        },
+        {
+            "state": {
+                "status": "restoring",
+                "progress": {
+                    "quantity": 70,
+                    "unit": "percent"
+                }
+            },
+            "balance": {
+                "total": {
+                    "quantity": 16,
+                    "unit": "lovelace"
+                },
+                "available": {
+                    "quantity": 163,
+                    "unit": "lovelace"
+                }
+            },
+            "name": "⨌]&y𫠏lg>lR診CTOuvLwV4:G%u|*=cqaBdA)26@P;.LM[=Zn(uO3/!z7X=~-@㨴5XFF$'zEv1pkh⨱)=𣟀[qcaw\\hpq@2_^K''Q;`嫦(𣑇Fa?u&FjC8𪟼Y|)?𦓖3OJK𢬌SqfK-`xi%mZm{9ysz}t3𫞟`Q8-",
+            "id": "22d0fd5a156f9c369825134217836e401ea775ac"
+        },
+        {
+            "passphrase": {
+                "last_updated_at": "2312-03-03T22:07:19Z"
+            },
+            "state": {
+                "status": "restoring",
+                "progress": {
+                    "quantity": 77,
+                    "unit": "percent"
+                }
+            },
+            "balance": {
+                "total": {
+                    "quantity": 55,
+                    "unit": "lovelace"
+                },
+                "available": {
+                    "quantity": 241,
+                    "unit": "lovelace"
+                }
+            },
+            "name": "Lv@~Fe$^Z=<dXoUgF^'8fXE}6W\\?(~o:L$?O-yI⅒2]椏巎SQ&KMhhxx\"L2ꗛfpi:0d]E=>Q)~DO𢤙K}4w좬uM3Cs,𡍠vꅻA~R;{:)OUm[𨎶~'3r6qcEk_@eXV^IY\\*?[𑒻MugsG𑈇;k`F^t6@S(Od'[b]V<.iD~@|[ghC~\"+Q8?{}m8 |'Q[d\\ER=\"D`Q|T~{}#z\\uo:8t|fUs&Km%E`>~🈢8'{9\"ꠋAdQ|o#TbL^TN?n B|G",
+            "id": "d5c3584e7d3a1db1fc9a1bd9af2455b2a1019ab2"
+        }
+    ]
+}
