+++ lib/http-bridge/test/integration/Cardano/WalletSpec.hs
@@ -13,9 +13,7 @@ import Cardano.BM.Trace
 import Cardano.Launcher
     ( StdStream (..) )
 import Cardano.Wallet
-    ( newWalletLayer )
-import Cardano.Wallet.DB
-    ( DBFactory (..) )
+    ( WalletLayer (..) )
 import Cardano.Wallet.HttpBridge.Compatibility
     ( HttpBridge, block0, byronBlockchainParameters )
 import Cardano.Wallet.HttpBridge.Environment
@@ -85,12 +83,6 @@ spec = do
             db <- MVar.newDBLayer
             let tl = HttpBridge.newTransactionLayer @'Testnet @SeqKey
             let genesis = (block0, byronBlockchainParameters @'Testnet)
-            let dbFactory = DBFactory
-                    { withDatabase = const (\with -> with db)
-                    , removeDatabase = \_ -> pure ()
-                    }
-            wl <- newWalletLayer @_ @_ @(HttpBridge 'Testnet)
-                nullTracer genesis nl tl dbFactory []
-            cb wl
+            cb $ WalletLayer nullTracer genesis nl tl db
         Left e -> throwIO e
     cfg = HttpBridge.HttpBridgeConfig (Right Testnet) Nothing Nothing [] Inherit
